Warn user if the magnetometer's values are weird.
The application currently has no way to tell if one of the base sensors is wrong.
I'll add a check after, for now just warn the user in the log so that they know it cannot work.
Also, little derp on the version code, tag v1.11 should obviously have the
v1.11 version code in build.gradle