Bugfixes and attempts to use quaternions for the gyroscope.
Fixed support for API18-22 (TIL that one should never write code at 2a.m and make a release).
Added a warning for the user if the magnetic sensor returns weird/wrong values.
Added a way to fill a little bit more the fields of the Sensor class.
Now using the calculations of the gravity sensor to calculate the gyroscope, this allows it to work in any orientation and return the right values.