Fix for ConcurrentModificationException and already present sensors.
Added a way to determine wether the sensor is natively present or not. This was already checked when adding it but not in the hook when registering a new listener. This led to native sensors getting overriden leading to either an absence of values if the module doesn't work or to loss of precision.