Less jitter, better way to handle existing sensors, starting to work on cardboard.
The readings now take in to account the sensors's accuracy. If the values haven't changed of more than the sensor's accuracy, we don't update them. This reduces the amount of noise in the outputs of the virtual sensors. I think smoothing the inputs is as important as smoothing the outputs.

Added a better way to check that the sensors haven't already been added. Previous way was static, this should be better.

Started to work on cardboard hooks as this is used in many applications. Carboard uses the NDK so we have to replace the values at the SDK level in the library. It doesn't seem to work yet.