Cardboard hooking finally starting to work.
It's still random, it's not the correct values, but it's going in a good direction. Also removed the settings menu which didn't lead anywhere and confused people. Remove the gui field that checked if the theorical values where the same as the real values as it couldn't work. Probably some other changed that I'm forgetting right now. But heh, not dead!